Split example into individual files, add readme, apply most feedback
